🚚 rename: useTimer 커스텀 훅 분리 #40
